pyproject.toml
[tool.poetry]
name = "skud"
version = "0.1.0"
description = ""
authors = ["RomanL"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
gunicorn = "^20.1.0"
pillow = "^9.4.0"
webhook-listener = "^1.2.0"
djangorestframework = "^3.14.0"
django-rest-framework-roles = "^0.6"
rest-framework-roles = "^1.0.3"
django-api-admin = "^1.0.3"
django-restful-admin = "^1.1.3"
channels = "^4.0.0"
daphne = "^4.0.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
